+++ tests/TestApp/Controllers/HomeTestController.php
@@ -6,6 +6,7 @@ use Busarm\PhpMini\App;
 use Busarm\PhpMini\Interfaces\ResponseInterface;
 
 use function Busarm\PhpMini\Helpers\log_debug;
+use function Busarm\PhpMini\Helpers\out;
 
 /**
  * PHP Mini Framework